fix parent animate to children, and move some transition to helpers
