+++ src/victory-util/transition-helpers.js
@@ -39,7 +39,7 @@ function getKeyedDataDifference(a, b) {
  *                           exiting[datum.key] will be true if the data is
  *                           exiting, and similarly for `entering`.
  */
-export function getNodeTransitions(oldData, nextData) {
+function getNodeTransitions(oldData, nextData) {
   const oldDataKeyed = oldData && getKeyedData(oldData);
   const nextDataKeyed = nextData && getKeyedData(nextData);
 
@@ -84,4 +84,83 @@ export function getDomainFromChildren(props, axis) {
     return childDomains.length === 0 ?
       [0, 1] : [Math.min(...childDomains), Math.max(...childDomains)];
   }
+}
+
+
+/**
+ * If a parent component has animation enabled, calculate the transitions
+ * for any data of any child component that supports data transitions
+ * Data transitions are defined as any two datasets where data nodes exist
+ * in the first set and not the second, in the second and not the first,
+ * or both.
+ *
+ * @param  {Children}  oldChildren   this.props.children from old props
+ * @param  {Children}  nextChildren  this.props.children from next props
+ *
+ * @return {Object}                  Object with the following properties:
+ *                                    - nodesWillEnter
+ *                                    - nodesWillExit
+ *                                    - childrenTransitions
+ *                                    - nodesShouldExit
+ *                                    - nodesShouldEnter
+ */
+
+export function getInitialTransitionState(oldChildren, nextChildren) {
+  let nodesWillExit = false;
+  let nodesWillEnter = false;
+
+  const getTransition = (oldChild, newChild) => {
+    if (!newChild || oldChild.type !== newChild.type) {
+      return {};
+    }
+
+    const { exiting, entering } =
+      getNodeTransitions(getChildData(oldChild), getChildData(newChild)) || {};
+
+    nodesWillEnter = nodesWillEnter || !!entering;
+    nodesWillExit = nodesWillExit || !!exiting;
+
+    return { entering: entering || false, exiting: exiting || false };
+  };
+
+  const getTransitionsFromChildren = (old, next) => {
+    return old.map((child, idx) => {
+      if (child.props.children) {
+        return getTransitionsFromChildren(
+          React.Children.toArray(old[idx].props.children),
+          React.Children.toArray(next[idx].props.children)
+        );
+      }
+      // get Transition exiting and enter nodes
+      return getTransition(child, next[idx]);
+    });
+  };
+
+  const childrenTransitions = getTransitionsFromChildren(
+    React.Children.toArray(oldChildren),
+    React.Children.toArray(nextChildren)
+  );
+
+  return {
+    childrenTransitions,
+    nodesWillExit,
+    nodesWillEnter,
+    // TODO: This may need to be refactored for the following situation.
+    //       The component receives new props, and the data provided
+    //       is a perfect match for the previous data and domain except
+    //       for new nodes. In this case, we wouldn't want a delay before
+    //       the new nodes appear.
+    nodesShouldExit: false,
+    nodesShouldEnter: false
+  };
+}
+
+
+export function checkContinuousChartType(child) {
+  const whitelistContinuous = ["line", "area"];
+
+  if (whitelistContinuous.indexOf(child.type.role) === -1) {
+    return false;
+  }
+  return true;
 }
